Revert "Finish refactoring out replay (#4850)"
This reverts commit a32cffa769c0a4470e6ebf6cd1da9293593c25d2.
